riskpooling/requirements_base.txt
otree-core>=1.4.17
Django==1.8.8 # for heroku, needs to be explicitly in requirements file
